get tasks from the json body instead of url query + improve docs

+++ backend/api/tasks/resources.py
@@ -134,47 +134,60 @@ class TasksQueriesJsonAPI(Resource):
     @token_auth.login_required
     def delete(self, project_id):
         """
-        Deletes all tasks from a project
+        Delete a list of tasks from a project
         ---
         tags:
             - tasks
         produces:
             - application/json
         parameters:
+            - in: header
+              name: Authorization
+              description: Base64 encoded session token
+              required: true
+              type: string
+              default: Token sessionTokenHere==
             - name: project_id
               in: path
               description: Project ID the task is associated with
               required: true
               type: integer
               default: 1
-            - in: query
-              name: tasks
-              type: string
-              description: List of tasks;
-              default: 1,2
+            - in: body
+              name: body
+              required: true
+              description: JSON object with a list of tasks to delete
+              schema:
+                  properties:
+                      tasks:
+                          type: array
+                          items:
+                              type: integer
+                          default: [ 1, 2 ]
         responses:
             200:
-                description: Project found
+                description: Task(s) deleted
             400:
                 description: Bad request
             403:
                 description: Forbidden
             404:
-                description: Project not found
+                description: Project or Task Not Found
             500:
                 description: Internal Server Error
         """
         user_id = token_auth.current_user()
         user = UserService.get_user_by_id(user_id)
         if user.role != UserRole.ADMIN.value:
-            return {"Error": "User is not admin"}, 403
+            return {"Error": "This endpoint action is restricted to ADMIN users."}, 403
 
-        tasks_ids = request.args.get("tasks")
+        tasks_ids = request.get_json().get("tasks")
         if tasks_ids is None:
             return {"Error": "Tasks ids not provided"}, 400
+        if type(tasks_ids) != list:
+            return {"Error": "Tasks were not provided as a list"}, 400
 
         try:
-            tasks_ids = [int(t) for t in tasks_ids.split(",")]
             ProjectService.delete_tasks(project_id, tasks_ids)
             return {"Success": "Task(s) deleted"}, 200
         except NotFound as e:
